feat(cli): deploy self-hosted langgraph agent server
