+++ packages/openui-cli/src/templates/openui-cloud/backends/langgraph/manifest.json
@@ -13,8 +13,9 @@
   "scripts": {
     "dev": "run-p dev:langgraph dev:next",
     "dev:langgraph": "langgraphjs dev",
-    "dev:next": "next dev"
+    "dev:next": "next dev",
+    "deploy:langgraph": "langgraphjs deploy"
   },
   "removeFiles": ["src/lib/tool-loop.ts"],
-  "gettingStarted": "The generated LangGraph agent uses OpenUI Cloud as its Responses provider. `{{packageManager}} run dev` starts both the Agent Server and Next.js; deploy `langgraph.json` to LangSmith and point LANGGRAPH_API_URL at it in your frontend deployment.\nAsk \"What's the weather in Berlin?\" to exercise the included LangGraph tool."
+  "gettingStarted": "The generated LangGraph agent uses OpenUI Cloud as its Responses provider. `{{packageManager}} run dev` starts both the Agent Server and Next.js; `{{packageManager}} run deploy:langgraph` deploys only the backend. Point LANGGRAPH_API_URL at it in your frontend deployment.\nAsk \"What's the weather in Berlin?\" to exercise the included LangGraph tool."
 }
