+++ packages/openui-cli/src/templates/openui-cloud/package.json
@@ -35,6 +35,7 @@
   "pnpm": {
     "onlyBuiltDependencies": [
       "@openuidev/lang-core",
+      "esbuild",
       "sharp",
       "unrs-resolver"
     ]
