+++ packages/openui-cli/src/templates/openui-cloud/pnpm-workspace.yaml
@@ -4,10 +4,12 @@
 # onlyBuiltDependencies here (10.14-10.x), allowBuilds here (>=11).
 onlyBuiltDependencies:
   - "@openuidev/lang-core"
+  - esbuild
   - sharp
   - unrs-resolver
 allowBuilds:
   "@openuidev/lang-core": true
+  esbuild: true
   sharp: true
   unrs-resolver: true
 packages: ["."]
