+++ packages/openui-cli/src/templates/openui-self-hosted/backends/langgraph/langgraph.json
@@ -1,0 +1,8 @@
+{
+  "node_version": "20",
+  "dependencies": ["."],
+  "graphs": {
+    "agent": "./src/agent/agent.ts:graph"
+  },
+  "env": ".env"
+}
