+++ packages/openui-cli/src/templates/openui-self-hosted/backends/langgraph/manifest.json
@@ -3,8 +3,20 @@
     "@langchain/core": "^1.2.5",
     "@langchain/langgraph": "^1.4.9",
     "@langchain/openai": "^1.5.6",
+    "@openuidev/langchain": "latest",
+    "langchain": "^1.5.5",
     "zod": "^4.4.3"
   },
+  "devDependencies": {
+    "@langchain/langgraph-cli": "^1.4.4",
+    "npm-run-all2": "^9.0.2"
+  },
+  "scripts": {
+    "dev": "run-p dev:langgraph dev:next",
+    "dev:langgraph": "langgraphjs dev",
+    "dev:next": "next dev",
+    "deploy:langgraph": "langgraphjs deploy"
+  },
   "removeDependencies": ["openai"],
-  "gettingStarted": "The generated API route uses LangGraph.\nAsk \"What's the weather in Berlin?\" to exercise its native tool loop."
+  "gettingStarted": "The generated LangGraph backend is a standalone Agent Server. `{{packageManager}} run dev` starts both the Agent Server and Next.js; `{{packageManager}} run deploy:langgraph` deploys only the backend. Point LANGGRAPH_API_URL at it in the frontend deployment.\nAsk \"What's the weather in Berlin?\" to exercise its native tool loop."
 }
