+++ packages/openui-cli/src/templates/openui-self-hosted/package.json
@@ -5,6 +5,7 @@
   "pnpm": {
     "onlyBuiltDependencies": [
       "@openuidev/lang-core",
+      "esbuild",
       "sharp",
       "unrs-resolver"
     ]
