+++ packages/openui-cli/src/templates/openui-self-hosted/pnpm-workspace.yaml
@@ -7,10 +7,12 @@ packages:
 # onlyBuiltDependencies here (10.14-10.x), allowBuilds here (>=11).
 onlyBuiltDependencies:
   - "@openuidev/lang-core"
+  - esbuild
   - sharp
   - unrs-resolver
 allowBuilds:
   "@openuidev/lang-core": true
+  esbuild: true
   sharp: true
   unrs-resolver: true
 minimumReleaseAgeExclude:
